feat: move onBrushEnd from config to Settings
