+++ packages/charts/src/chart_types/heatmap/state/selectors/on_brush_end_caller.ts
@@ -12,8 +12,8 @@ import { ChartType } from '../../..';
 import { GlobalChartState } from '../../../../state/chart_state';
 import { createCustomCachedSelector } from '../../../../state/create_selector';
 import { getLastDragSelector } from '../../../../state/selectors/get_last_drag';
+import { getSettingsSpecSelector } from '../../../../state/selectors/get_settings_specs';
 import { DragCheckProps, hasDragged } from '../../../../utils/events';
-import { getHeatmapConfigSelector } from './get_heatmap_config';
 import { getPickedCells } from './get_picked_cells';
 import { getSpecOrNull } from './heatmap_spec';
 import { isBrushEndProvided } from './is_brush_available';
@@ -35,7 +35,7 @@ export function createOnBrushEndCaller(): (state: GlobalChartState) => void {
         return;
       }
       selector = createCustomCachedSelector(
-        [getLastDragSelector, getSpecOrNull, getHeatmapConfigSelector, getPickedCells],
+        [getLastDragSelector, getSpecOrNull, getSettingsSpecSelector, getPickedCells],
         (lastDrag, spec, { onBrushEnd }, pickedCells): void => {
           const nextProps: DragCheckProps = {
             lastDrag,
